# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing venders",
        "title": "Vender API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/vender": {
            "get": {
                "description": "Fetch all vender profiles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vender"
                ],
                "summary": "Get all venders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.VenderProfile"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new vender profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vender"
                ],
                "summary": "Create a new vender",
                "parameters": [
                    {
                        "description": "Vender Data",
                        "name": "vender",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VenderProfile"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vender/{vender_id}": {
            "get": {
                "description": "Fetch a vender using vender_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vender"
                ],
                "summary": "Get vender by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vender ID",
                        "name": "vender_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VenderProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update vender details by vender_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vender"
                ],
                "summary": "Update vender profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vender ID",
                        "name": "vender_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated vender data",
                        "name": "vender",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VenderProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "enum.VenderStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE",
                "SUSPENDED"
            ],
            "x-enum-varnames": [
                "Active",
                "Inactive",
                "Suspended"
            ]
        },
        "enum.VenderType": {
            "type": "string",
            "enum": [
                "INDIVISUAL",
                "CLINIC",
                "HOSPITAL"
            ],
            "x-enum-varnames": [
                "Indivisual",
                "Clinic",
                "Hospital"
            ]
        },
...
        "model.VenderProfile": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "phone_number",
                "status",
                "vender_type"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone_number": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "INACTIVE",
                        "SUSPENDED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.VenderStatus"
                        }
                    ]
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vender_id": {
                    "type": "string"
                },
                "vender_type": {
                    "enum": [
                        "INDIVISUAL",
                        "CLINIC",
                        "HOSPITAL"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.VenderType"
                        }
                    ]
                }
            }
        }
    }
}